setup for serverside implementation of batch import
