+++ src/services/user/hooks/batch-import.js
@@ -1,0 +1,26 @@
+'use strict';
+
+const globalHooks = require('../../../hooks');
+const hooks = require('feathers-hooks');
+const auth = require('feathers-authentication');
+const errors = require('feathers-errors');
+
+exports.before = {
+    all: [auth.hooks.authenticate('jwt')],
+    find: [],
+    get: [],
+    create: [],
+    update: [],
+    patch: [],
+    remove: []
+}
+
+exports.after = {
+    all: [],
+    find: [],
+    get: [],
+    create: [],
+    update: [],
+    patch: [],
+    remove: []
+}
