+++ src/services/user/index.js
@@ -1,30 +1,13 @@
 'use strict';
 
-const service = require('feathers-mongoose');
-const user = require('./model');
-const hooks = require('./hooks');
+const modelService = require('./model-service');
+const batchImportService = require('./batch-import-service');
 
 module.exports = function () {
-	const app = this;
+    const app = this;
 
-	const options = {
-		Model: user,
-		paginate: {
-			default: 25,
-			max: 1000
-		},
-		lean: true
-	};
-
-	// Initialize our service with any options it requires
-	app.use('/users', service(options));
-
-	// Get our initialize service to that we can bind hooks
-	const userService = app.service('/users');
-
-	// Set up our before hooks
-	userService.before(hooks.before(app));	// TODO: refactor
-
-	// Set up our after hooks
-	userService.after(hooks.after);
+    //setup model services
+    app.configure(modelService);
+    //setup batch import services
+    app.configure(batchImportService);
 };
